+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/AbstractVisitorRule.java
@@ -2,38 +2,31 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.ast.impl.antlr4;
+package net.sourceforge.pmd.lang.rule;
 
 import net.sourceforge.pmd.RuleContext;
 import net.sourceforge.pmd.lang.ast.AstVisitor;
 import net.sourceforge.pmd.lang.ast.Node;
-import net.sourceforge.pmd.lang.rule.AbstractRule;
 
-/**
- * Base implementation of an antlr rule.
- */
-public abstract class AntlrBaseRule extends AbstractRule {
-
-    protected AntlrBaseRule() {
-        // inheritance constructor
-    }
-
+public abstract class AbstractVisitorRule extends AbstractRule {
     @Override
     public void apply(Node target, RuleContext ctx) {
         AstVisitor<RuleContext, ?> visitor = buildVisitor();
         assert visitor != null : "Rule should provide a non-null visitor";
-        assert target instanceof AntlrNode : "Incorrect node type " + target + " passed to " + this;
 
-        ((AntlrNode<?>) target).acceptVisitor(visitor, ctx);
+        target.acceptVisitor(visitor, ctx);
     }
 
     /**
      * Returns a rule visitor that can visit nodes for the given rule context.
      * This visitor should explore the nodes it's interested in and report
      * violations on the given rule context.
+     * <p>
+     *     Language specific subclasses should redefine the return type to use
+     *     a language specific visitor interface.
+     * </p>
      *
      * @return A visitor bound to the given rule context
      */
     public abstract AstVisitor<RuleContext, ?> buildVisitor();
-
 }