+++ docs/pages/release_notes.md
@@ -43,6 +43,7 @@ The remaining section describe the complete release notes for 7.0.0.
 * Moved the two classes {% jdoc core::cpd.impl.AntlrTokenizer %} and {% jdoc core::cpd.impl.JavaCCTokenizer %} from
   `internal` package into package {% jdoc_package core::cpd.impl %}. These two classes are part of the API and
   are base classes for CPD language implementations.
+* `AntlrBaseRule` is gone in favor of {% jdoc core::pmd.lang.rule.AbstractVisitorRule %}.
 
 #### Fixed Issues:
 * java-codestyle